Use proper resolver logic for any RR type
Instead of copypasta for the DNS resolution logic, accept a tiny bit of
overhead, by using a callback for massaging out the RR data into an
`interface{}`, called from a generic routine.  Then have the A/AAAA
lookups use that, and massage the `[]interface{}` into an `[]net.IP`,
which should result in a noop behavior change for that routine.

Then switch the TLSA lookups to use this too.