Doc access more fully; fix bugs, add -nocertnames
Went through ensuring that the `Access needed` section was up-to-date.

Went with `-nocertnames` flag to inhibit loading, not a build flag.

Testing, identified a crash-bug when `DNS_RESOLVER` is in environment:
the recent work to add DNS retry logic assumed that we had the client
config built from reading `/etc/resolv.conf`, which we don't do if given
that variable.  So create a stub config.

(I'd redo the logic to let us put all the `DNS_RESOLVER`-based stuff in
 there, but the env var lets us override port per-server, not quite
 willing to shoe-horn that in.)