Change CNAME in Base Domain for TLSA lookup
RFC 7671 lays down rules for CNAME.

I'd beenin the mindset of MX host targets can't be CNAMEs, but we
support non-MX lookup (MX has to be explicitly enabled!) and people do
it anyway.

So address resolution finds and returns the resolved name too, we try
the TLSA-via-CNAME before TLSA-via-original and we throw a small snit if
this is for hosts from MX, by using a red message (but don't class it an
outright error).