Add -quiet with much re-jigging needed
Ability to batch output messages, to be able to drop them on the floor
and only emit them if errored; reused programStatus as the container.

Using close() as the signal for done, needed too-careful attention to
whether the children-exist waitGroup done in child go-routines or not,
so moved some `go foo()` calls to `fooGo()` where `fooGo()` does
bookkeeping around waitgroups before calling `go foo()`.

Added a `-debug` flag, to catch program flow.  The batched emitters thus
become a nice go-routine tracker as they exit.

Added `-q` as alias for `-quiet` and `-p` as alias for `-port`.

By the time I was done, the ugliness debt was probably a slight net win
because the probing waitgroup is a bit less visible.