BugFix: SRV lookup path never exited
The defer based on ChildBatcher was triggered immediately at the start
here, as for other functions, but I didn't first create the
ChildBatcher, leaving that until after the lookup string was parsed, to
have a nice label.  This was a bug, we deferred on the wrong object.

Defer on the correct object.  Handle the other exit path cleanly.