Add -submissions flag and rework doc a little
RFC8314 made submissions (aka smtps or ssmtp) official, with port 465
allowed etc, so we add a new `-submissions` flag which is just like
`-submission` but which also forces on the `-tls-on-connect` flag.

I made sure we could handle SMTPS before, but cowardly avoided the
headache of standards trolls in my previous wording.

Add `[port nnnn]` to the end of the TLS/STARTTLS waffle message, so
we can confirm which port we're actually talking on.