prep for go 1.13 minimum (errors.As change)
Go 1.15 changes `tls.Conn()` to return a _wrapped_ `net.Error` such that an
interface cast will no longer work; code now has to use `errors.As()`,
introduced with Go 1.13, to walk the chain to find an error which might satisfy
that interface.

We're not using `tls.Conn()` but I'm worried that this is a portent of things
to come, so I'm switching now.

This commit updates the README and changes the minimum version compile guards,
but does not change the error matching logic, which will be the next commit.