Switch from error type conversions to errors.As
The Go stdlib maintainers apparently feel that changing the top-level error
type does not break the compatibility guarantee, with such a change in Go 1.15.
Go 1.13 introduced a way to simply match, allowing for chaining, so switch from
checking if an error is a type, to seeing if an error in the chain is of that
type.