resolvers: avoid hitting all resolvers in ok case
There was a missing loop break for when we had finally decided we were happy
with a DNS response, to avoid hitting the next DNS resolver.

Use the `debugf()` facility to debug-log each time we're about to talk to a
resolver, under `-debug`.

Add a panic in case I've misunderstood what the me of 6 years ago was thinking
for this DNS retry loop.

For the scenario where AD is missing, be explicit that we're skipping the other
resolvers.  It took me a while to understand my prior reasoning, but it's
obvious in retrospect: missing AD is almost always going to be "zone is
unsigned", so there's no point trying the other resolvers.  For the rare
scenario where local DNS resolvers are broken ... we are not a debug-diagnosis
tool for local caching resolvers, we're a tool for querying state of remote
domains.