admin: updates.sh warn retracted in prior state
We also need to know if the _prior_ state depended upon modules which have now
been retracted, so that we can investigate why they were retracted and if that
in turn requires us to take action.

This should help with supply-chain attacks, somewhat.

Open question as to whether or not we should do the same with zizmor
before/after the pinact checks.